move style components
