+++ src/styles/SelectorStyle.js
